Admin2

+++ app/Http/Controllers/UserController.php
@@ -12,6 +12,7 @@ use App\Http\Requests\UserRequest;
 use App\Http\Requests\SetPasswordRequest;
 use App\Http\Requests\AccountPasswordRequest;
 use App\Http\Requests\WishlistRequest;
+use App\Http\Requests\FriendRequest;
 use App\Wishlist;
 use App\User;
 use App\DefaultWishlist;
@@ -82,9 +83,10 @@ class UserController extends Controller
   {
     $user = Auth::user();
 
-   $requests = Friend::with('friendRequest')->where('status', '=', '0')->get();
-   // dd($requests);
-   // die();
+    // $requests = Friend::with('friendRequest')->where('status', '=', '0')->get();
+    $requests = Friend::where('friend_userid', '=', $user->id)->where('status', '=', '0')->get();
+    // print($requests);
+    // die();
 
    return view('userlayouts.notifications', compact('requests'));
   }
@@ -143,39 +145,50 @@ class UserController extends Controller
     if($user->id != $id){
       $otherUser = User::where('id', '=', $id)->firstorFail();
 
-      // $friend = Friend::with('user')
-      //                   ->where('userid', '=', $user->id)
-      //                   ->where('friend_userid', '=', $otherUser->id)
-      //                   ->first();
+      $requests = Friend::with('friendRequest')
+                          ->where('userid', '=', $otherUser->id)
+                          ->where('friend_userid', '=', $userId)
+                          ->where('status', '=', 0)
+                          ->get();
+      // print($requests);
+      // die();
 
       $usersWithFriends = User::with('friendsOfMine', 'friendOf')->get();
       $friends = User::find($otherUser->id)->friends;
       // print($friend);
       // die();
-      if(count($friends)==0){
-          $friendRequest = Friend::where('userid', '=', $user->id)
-                                ->where('friend_userid', '=', $id)
-                                ->first();
+      // if(count($friends)==0){
+      $friendRequest = Friend::where('userid', '=', $user->id)
+                            ->where('friend_userid', '=', $id)
+                            ->where('status', '=', 0)
+                            ->first();
 
-          if(!empty($friendRequest))
-            $status = 0;
+      if(!empty($friendRequest)){
+        $status = 0;
       }
-      else{
+      else {
+        $status = 3;
+      }
+
+      if(!empty($friends)){
         foreach ($friends as $f) {
           if ($f->id == $userId) {
             $status = $f->pivot->status;
           }
         }
-
+      }
+      else{
+        $status= 0;
       }
 
 
 
-      if($otherUser->privacy == 0){
-          return view('otheruser.otheruserprofile', compact('otherUser', 'friends', 'status'));
+      if(($otherUser->privacy == 1) && $status != 1){
+
+          return view('otheruser.otheruserprivate', compact('otherUser', 'friends', 'status', 'requests'));
       }
       else {
-          return view('otheruser.otheruserprivate', compact('otherUser', 'friends', 'status'));
+          return view('otheruser.otheruserprofile', compact('otherUser', 'friends', 'status', 'requests'));
       }
     }
     else {
@@ -325,13 +338,14 @@ class UserController extends Controller
     $user = Auth::user();
     $id = $user->id;
     $newImage = '';
+    $hostURL = '192.168.1.17';
     $newImage = Input::file('imageurl');
     $filename  = $user->id . time() . '.' . $newImage->getClientOriginalExtension();
     // dd($filename);
 
-    $path = public_path('img/userImages/' . $filename);
+    $path = ('C:/xampp/htdocs/wishareimages/userimages/' . $filename);
     Image::make($newImage->getRealPath())->fit(150, 150)->save($path);
-    $user->imageurl = 'img/userImages/'.$filename;
+    $user->imageurl =  'http://' . $hostURL . '/wishareimages/userimages/'.$filename;
 
     $user->save();
 
@@ -416,22 +430,28 @@ class UserController extends Controller
                         ->orderBy('created_at', 'desc')
                         ->get();
     if(count($wishlists) > 0 || !empty($user))
-      return view('userlayouts.profile', compact('user', 'wishlists'));
+      return view('userlayouts.wishlistProfile', compact('user', 'wishlists'));
   }
 
   public function addFriend($id)
   {
     $user = Auth::user();
 
-    $friend = new Friend(array(
-      'friend_userid' => $id,
-      'userid' => $user->id,
-      'date_added' => date("Y-m-d h:i:s"),
-      'status' => 0,
-      'seen' => 0,
-    ));
-    // dd($friend);
-    $friend->save();
+    $exists = Friend::where('friend_userid', '=', $id)
+                      ->where('userid', '='. $user->id)
+                      ->get();
+    // dd($exists);
+    if(!empty($exists)){
+      $friend = new Friend(array(
+        'friend_userid' => $id,
+        'userid' => $user->id,
+        'date_added' => date("Y-m-d h:i:s"),
+        'status' => 0,
+        'seen' => 0,
+      ));
+      // dd($friend);
+      $friend->save();
+    }
 
     return redirect()->action('UserController@otheruser', [$id]);
 
@@ -444,13 +464,16 @@ class UserController extends Controller
 
     $usersWithFriends = User::with('friendsOfMine', 'friendOf')->get();
     $friend = User::find($user->id)->friends;
+    // print($friend);
+    // die();
 
-    foreach ($friend as $f) {
-      if ($f->id == $id) {
-        Friend::destroy($f->pivot->id);
+    if(!empty($friend)){
+      foreach ($friend as $f) {
+        if ($f->id == $id) {
+          Friend::destroy($f->pivot->id);
+        }
       }
     }
-
 
     // $friend->delete();
 
@@ -464,10 +487,12 @@ class UserController extends Controller
 
     $friendRequest = Friend::where('userid', '=', $user->id)
                             ->where('friend_userid', '=', $id)
-                            ->firstorFail();
-    // var_dump($friendRequest);
-
-    $friendRequest->delete();
+                            ->where('status', '=', 0)
+                            ->first();
+    // print($friendRequest);
+    // die();
+    if(!empty($friendRequest))
+      $friendRequest->delete();
 
     return redirect()->action('UserController@otheruser', [$id]);
 
@@ -475,22 +500,35 @@ class UserController extends Controller
 
   public function acceptFriendRequest($id)
   {
-    $friendRequest = Friend::find($id)->first();
+    $user = Auth::user();
 
-    $friendRequest->date_accepted = date("Y-m-d h:i:s");
-    $friendRequest->status = 1;
+    $friendRequest = Friend::where('id', '=', $id)
+                            ->where('friend_userid','=', $user->id)
+                            ->where('status', '=', 0)
+                            ->first();
+    // print($friendRequest);
+    // die();
+    if(!empty($friendRequest)){
+      $friendRequest->date_accepted = date("Y-m-d h:i:s");
+      $friendRequest->status = 1;
 
-    $friendRequest->save();
+      $friendRequest->save();
+    }
 
-    // dd($friendRequest);
     return redirect()->action('UserController@notifications');
   }
 
   public function declineFriendRequest($id)
   {
-    $friendRequest = Friend::find($id)->first();
+    $user = Auth::user();
 
-    $friendRequest->delete();
+    $friendRequest = Friend::find($id)
+                    ->where('friend_userid','=', $user->id)
+                    ->where('status', '=', 0)
+                    ->first();
+
+    if(!empty($friendRequest))
+      $friendRequest->delete();
 
     // dd($friendRequest);
     return redirect()->action('UserController@notifications');
